Tests/LibWeb/Text/expected/canvas/invalid-font-property.txt

20px SerenitySans
20px SerenitySans
20px SerenitySans
20px SerenitySans
